fix: Improve run task node ID calculation.

+++ crates/action-pipeline/src/job_context.rs
@@ -26,7 +26,7 @@ pub struct JobContext {
     pub result_sender: Sender<Action>,
 
     /// Currently running jobs (used by the dispatcher)
-    pub running_jobs: Arc<RwLock<FxHashMap<NodeIndex, u32>>>,
+    pub running_jobs: Arc<RwLock<FxHashMap<NodeIndex, u64>>>,
 
     /// Acquires a permit for concurrency
     pub semaphore: Arc<Semaphore>,
